move InspectRandomNodeHelper test cases to ContractFacade
